+++ app/assets/stylesheets/components/banner-editor.scss
@@ -1,0 +1,39 @@
+
+.banner-logo-label:hover {
+  background: linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5));
+}
+
+.banner-background-image-label:hover {
+  background: linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5));
+}
+
+.banner-headline:focus{
+ box-shadow: 0 0 2pt 2pt #7d7bdc;
+ outline: none;
+}
+
+.banner-headline:hover{
+ box-shadow: 0 0 2pt 2pt #7d7bdc;
+}
+
+.banner-description:focus{
+ box-shadow: 0 0 2pt 2pt #7d7bdc;
+ outline: none;
+}
+
+.banner-description:hover{
+ box-shadow: 0 0 2pt 2pt #7d7bdc;
+}
+
+.banner-social-link-input:focus{
+ box-shadow: 0 0 2pt 2pt #7d7bdc;
+ outline: none;
+}
+
+.banner-social-link-input:hover{
+ box-shadow: 0 0 2pt 2pt #7d7bdc;
+}
+
+.banner-link-option:hover{
+  background-color: rgb(241, 241, 249);
+}
