+++ app/assets/stylesheets/pages/banner-intro.scss
@@ -1,0 +1,24 @@
+#instant-donation-modal-selection {
+  padding: 10px 40px 10px 40px;
+}
+
+#icn-donation-jar {
+  height: 95px;
+}
+
+.banner-intro-headline {
+  display: inline;
+}
+
+.banner-intro-tagline {
+  display: inline;
+  padding-left: 10px;
+}
+
+.banner-intro-intro {
+  padding-top: 20px;
+}
+
+#open-banner-button {
+  margin-top: 20px;
+}
