--- app/javascript/packs/brave_rewards_banner.jsx
@@ -1,457 +1,0 @@
-import React from 'react'
-
-import BatsBackground from '../../assets/images/bg_bats.svg'
-import HeartsBackground from '../../assets/images/bg_hearts.svg'
-import { BatColorIcon, YoutubeColorIcon, TwitterColorIcon, TwitchColorIcon } from 'brave-ui/components/icons'
-import Checkbox from 'brave-ui/components/formControls/checkbox'
-
-import {styles} from '../packs/brave_rewards_banner.style.jsx'
-
-export default class BraveRewardsBanner extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      title: 'Your title',
-      description: 'Welcome to Brave Rewards banner',
-      backgroundImage: null,
-      backgroundImageData: '',
-      logoImage: null,
-      logoImageData: '',
-      youtube: '',
-      twitter: '',
-      twitch: '',
-      donationAmounts: [1, 5, 10],
-      mode: 'Edit',
-      width: '1320'
-    }
-     this.handleBackgroundImageUpload = this.handleBackgroundImageUpload.bind(this);
-     this.handleLogoImageUpload = this.handleLogoImageUpload.bind(this);
-     this.updateTitle = this.updateTitle.bind(this);
-     this.updateDescription = this.updateDescription.bind(this)
-     this.handleSave = this.handleSave.bind(this);
-     this.fetchSiteBanner = this.fetchSiteBanner.bind(this);
-     this.updateYoutube = this.updateYoutube.bind(this);
-     this.updateTwitter = this.updateTwitter.bind(this);
-     this.updateTwitch = this.updateTwitch.bind(this);
-     this.fetchSiteBanner = this.fetchSiteBanner.bind(this);
-     this.updateWindowDimensions = this.updateWindowDimensions.bind(this);
-  }
-
-  componentWillMount(){
-    this.modalize();
-  }
-
-  componentDidMount(){
-    this.fetchSiteBanner();
-    document.getElementsByClassName('brave-rewards-banner-control-bar-save-button')[0].addEventListener("click", this.handleSave);
-    window.addEventListener('resize', this.updateWindowDimensions);
-  }
-
-  updateWindowDimensions() {
-    if(window.innerWidth < 991){
-      this.close();
-    }
-    else{
-      this.setState({ width: 1320});
-    }
-}
-
-  modalize(){
-    document.getElementsByClassName("modal-panel")[0].style.maxWidth = 'none';
-    document.getElementsByClassName("modal-panel")[0].style.padding = '0px';
-    document.getElementsByClassName("modal-panel--content")[0].style.padding = '0px';
-  }
-
-  close(){
-    document.getElementsByClassName("modal-panel")[0].style.maxWidth = '40rem';
-    document.getElementsByClassName("modal-panel")[0].style.padding = '2rem 2rem';
-    document.getElementsByClassName("modal-panel--content")[0].style.padding = '1rem 1rem 0 1rem';
-    document.getElementsByClassName("modal-panel--close js-deny")[0].click();
-  }
-
-  fetchSiteBanner(){
-    let that = this
-    let id = document.getElementById("publisher_id").value;
-    let url = '/publishers/' + id + "/site_banners/fetch";
-
-    fetch(url, {
-      method: 'GET',
-      headers: {
-        'Accept': 'text/html',
-        'X-Requested-With': 'XMLHttpRequest',
-        'X-CSRF-Token': document.head.querySelector("[name=csrf-token]").content
-        },
-        credentials: "same-origin",
-        }).then(function(response) {
-          return response.json();
-        })
-        .then(function(banner) {
-
-          that.setState({
-            title: banner.title,
-            description: banner.description,
-            youtube: banner.social_links.youtube,
-            twitter: banner.social_links.twitter,
-            twitch: banner.social_links.twitch,
-            donationAmounts: banner.donation_amounts,
-          })
-
-          that.cropFetchedLogo(banner.logoImage, that);
-          that.cropFetchedBackgroundImage(banner.backgroundImage, that)
-        });
-  }
-
-  updateTitle(event){
-     this.setState({title: event.target.value})
-  }
-
-  updateDescription(event){
-     this.setState({description: event.target.value})
-  }
-
-  updateYoutube(event){
-     this.setState({youtube: event.target.value})
-  }
-
-  updateTwitter(event){
-     this.setState({twitter: event.target.value})
-  }
-
-  updateTwitch(event){
-     this.setState({twitch: event.target.value})
-  }
-
-  updateDonationAmounts(event, index){
-    let temp = this.state.donationAmounts
-    if(/^(\s*|\d+)$/.test(event.target.value)){
-      temp[index] = event.target.value
-      this.setState({donationAmounts: temp})
-    }
-  }
-
-  handleBackgroundImageUpload(event) {
-    let that = this;
-    this.cropBackgroundImage(event, that);
-  }
-
-  handleLogoImageUpload(event) {
-    let that = this;
-    this.cropLogo(event, that);
-  }
-
-  cropFetchedBackgroundImage(backgroundImage, that){
-    let img = new Image();
-    img.src = backgroundImage;
-    img.onload = function() {
-      var canvas = document.createElement('canvas');
-      var ctx = canvas.getContext('2d');
-      canvas.width = 900;
-      canvas.height = 176;
-      ctx.drawImage(img, 0, 0, canvas.width, canvas.height);
-      let url = canvas.toDataURL('image/jpeg', 1);
-      that.setState({backgroundImage: url});
-    }
-  }
-
-  cropBackgroundImage(event, that){
-    if (event.target.files && event.target.files[0]) {
-      var filerdr = new FileReader();
-
-      filerdr.onload = function(e) {
-        var img900 = new Image();
-
-      img900.onload = function() {
-        var canvas = document.createElement('canvas');
-        var ctx = canvas.getContext('2d');
-        canvas.width = 900;
-        canvas.height = 176;
-        ctx.drawImage(img900, 0, 0, canvas.width, canvas.height);
-        let url = canvas.toDataURL('image/jpeg', 1);
-        that.setState({backgroundImage: url});
-        canvas.toBlob(function(blob){
-          let file = {};
-          file["files"] = [blob]
-          that.setState({backgroundImageData: file});
-        });
-      }
-      img900.src = e.target.result;
-    }
-    filerdr.readAsDataURL(event.target.files[0]);
-  }
-  }
-
-  cropFetchedLogo(logo, that){
-    let img = new Image();
-    img.src = logo;
-    img.onload = function() {
-      var canvas = document.createElement('canvas');
-      var ctx = canvas.getContext('2d');
-      canvas.width = 160;
-      canvas.height = 160;
-      ctx.drawImage(img, 0, 0, canvas.width, canvas.height);
-      let url = canvas.toDataURL('image/jpeg', 1);
-      that.setState({logoImage: url});
-    }
-  }
-
-  cropLogo(event, that){
-    if (event.target.files && event.target.files[0]) {
-      var filerdr = new FileReader();
-
-      filerdr.onload = function(e) {
-        var img160 = new Image();
-        var img480 = new Image();
-
-      img480.onload = function() {
-        var canvas = document.createElement('canvas');
-        var ctx = canvas.getContext('2d');
-        canvas.width = 480;
-        canvas.height = 480;
-        ctx.drawImage(img480, 0, 0, canvas.width, canvas.height);
-        canvas.toBlob(function(blob){
-          let file = {};
-          file["files"] = [blob]
-          that.setState({logoImageData: file});
-        });
-      }
-
-      img160.onload = function() {
-        var canvas = document.createElement('canvas');
-        var ctx = canvas.getContext('2d');
-        canvas.width = 160;
-        canvas.height = 160;
-        ctx.drawImage(img160, 0, 0, canvas.width, canvas.height);
-        let url = canvas.toDataURL('image/jpeg', 1);
-        that.setState({logoImage: url});
-      }
-
-      img160.src = e.target.result;
-      img480.src = e.target.result;
-    }
-    filerdr.readAsDataURL(event.target.files[0]);
-  }
-  }
-
-  handleSave(event) {
-    let that = this
-    let id = document.getElementById("publisher_id").value;
-    console.log(id);
-    let url = '/publishers/' + id + "/site_banners";
-    let body = new FormData();
-
-    body.append('title', this.state.title);
-    body.append('description', this.state.description);
-    body.append('donation_amounts', JSON.stringify(this.state.donationAmounts));
-    body.append('social_links', JSON.stringify({youtube: this.state.youtube, twitter: this.state.twitter, twitch: this.state.twitch}));
-
-    fetch(url, {
-      method: 'POST',
-      headers: {
-            'Accept': 'text/html',
-            'X-Requested-With': 'XMLHttpRequest',
-            'X-CSRF-Token': document.head.querySelector("[name=csrf-token]").content
-          },
-          credentials: "same-origin",
-          body: body
-    }).then (
-      function(response) {
-        function submitById(id, suffix) {
-
-        const url = '/publishers/' + document.getElementById("publisher_id").value + "/site_banners/update_" + suffix;
-
-        let file;
-        if(suffix === 'background_image'){ file = that.state.backgroundImageData }
-        else if(suffix === 'logo'){ file = that.state.logoImageData }
-
-        if (file === "" || file === null) { return; }
-          var reader = new FileReader();
-          reader.readAsDataURL(file.files[0]);
-
-          reader.onloadend = function () {
-            const body = new FormData();
-            body.append('image', reader.result);
-            fetch(url, {
-              method: 'POST',
-              headers: {
-                'Accept': 'text/html',
-                'X-Requested-With': 'XMLHttpRequest',
-                'X-CSRF-Token': document.head.querySelector("[name=csrf-token]").content
-              },
-              credentials: "same-origin",
-              body: body
-            });
-          }
-        }
-
-        if (response.status === 200) {
-          submitById("background-image-select-input", "background_image");
-          submitById("logo-image-select-input", "logo");
-        }
-      }).then(
-        function(response) {
-          that.close();
-        });
-  }
-
-  render() {
-
-    let style = styles
-
-    let logoImg
-    let logoLabel
-    let backgroundImageLabel
-    let backgroundImg
-    let explanatoryTitle
-    let explanatoryDescription
-    let socialLinkText
-
-    let rewardsBanner = {
-      maxWidth: this.state.width,
-      height:'488px',
-      overflow:'hidden'
-    }
-
-    if(this.props.mode === 'Edit'){
-      logoLabel = { height:'100%', width:'100%', borderRadius: '50%', border: '2px dotted white', cursor:'pointer'}
-      backgroundImageLabel = {height:'100%', width:'100%', border: '2px dotted white', cursor:'pointer'}
-
-      explanatoryTitle = {width:'100%', height:'50px', backgroundColor: 'rgba(0, 0, 0, 0)', border: '1px solid lightGray', borderRadius: '4px', marginTop: '15px', fontSize: '32px', color: '#686978'}
-      explanatoryDescription = {width:'100%', height:'150px', resize: 'none', backgroundColor: 'rgba(0, 0, 0, 0)', border:'1px solid lightGray', borderRadius: '4px', marginTop: '25px', fontSize: '22px', color: '#686978'}
-      socialLinkText = {marginTop:'auto', marginBottom:'auto', borderBottom: '1px solid lightGray', borderTop: '1px solid rgba(0, 0, 0, 0)', borderLeft: '1px solid rgba(0, 0, 0, 0)', color: '#686978', width: '90px', fontSize: '15px', backgroundColor: 'rgba(0, 0, 0, 0)', borderRadius: '0px'}
-
-      if(this.state.backgroundImage === null){
-        backgroundImg = {height: '176px', padding: '10px', background: `url(${BatsBackground}) left top no-repeat, url(${HeartsBackground}) right top no-repeat, rgb(158, 159, 171)`}
-      }
-      else{
-        backgroundImg = {height: '176px', padding: '10px', background: `linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)), url(${this.state.backgroundImage})`}
-      }
-      if(this.state.logoImage === null){
-        logoImg = {position: 'absolute', top: '190px', left: '35px', borderRadius: '50%', width: '160px', height: '160px', border: '6px solid white', padding: '10px', backgroundColor:'#FB542B'}
-      }
-      else{
-        logoImg = {position: 'absolute', top: '190px', left: '35px', borderRadius: '50%', width: '160px', height: '160px', border: '6px solid white', padding: '10px', background:`url(${this.state.logoImage})`}
-      }
-    }
-    else{
-      logoLabel = { height:'100%', width:'100%', borderRadius: '50%', border: 'none', pointerEvents: 'none'}
-      backgroundImageLabel = {height:'100%', width:'100%', border: 'none', pointerEvents: 'none'}
-
-      explanatoryTitle = {width:'100%', height:'50px', backgroundColor: 'rgba(0, 0, 0, 0)', border: '1px solid rgba(0, 0, 0, 0)', borderRadius: '4px', marginTop: '15px', fontSize: '32px', color: '#686978', userSelect:'none'}
-      explanatoryDescription = {width:'100%', height:'150px', resize: 'none', backgroundColor: 'rgba(0, 0, 0, 0)', border:'1px solid rgba(0, 0, 0, 0)', borderRadius: '4px', marginTop: '25px', fontSize: '22px', color: '#686978', userSelect:'none'}
-      socialLinkText = {marginTop:'auto', marginBottom:'auto', borderBottom: '1px solid rgba(0, 0, 0, 0)', borderTop: '1px solid rgba(0, 0, 0, 0)', borderLeft: '1px solid rgba(0, 0, 0, 0)', color: '#686978', width: '90px', fontSize: '15px', backgroundColor: 'rgba(0, 0, 0, 0)', borderRadius: '0px', userSelect: 'none'}
-
-      if(this.state.backgroundImage === null){
-        backgroundImg = {height: '176px', padding: '10px', background: `url(${BatsBackground}) left top no-repeat, url(${HeartsBackground}) right top no-repeat, rgb(158, 159, 171)`}
-      }
-      else{
-        backgroundImg = {height: '176px', padding: '10px', background: `url(${this.state.backgroundImage})`}
-      }
-      if(this.state.logoImage === null){
-        logoImg = {position: 'absolute', top: '190px', left: '35px', borderRadius: '50%', width: '160px', height: '160px', border: '6px solid white', padding: '10px', backgroundColor:'#FB542B'}
-      }
-      else{
-        logoImg = {position: 'absolute', top: '190px', left: '35px', borderRadius: '50%', width: '160px', height: '160px', border: '6px solid white', padding: '10px', background:`url(${this.state.logoImage})`}
-      }
-    }
-
-    style.logoLabel = logoLabel
-    style.backgroundImageLabel = backgroundImageLabel
-    style.logoImg = logoImg
-    style.backgroundImg = backgroundImg
-    style.explanatoryTitle = explanatoryTitle
-    style.explanatoryDescription = explanatoryDescription
-    style.socialLinkText = socialLinkText
-    style.rewardsBanner = rewardsBanner
-
-    return (
-      <div style={style.rewardsBanner} className="brave-rewards-banner">
-
-        <div className="brave-rewards-banner-logo" style={style.logoImg}>
-          <input type="file" id="logoImageInput" style={style.imageInput} onChange={this.handleLogoImageUpload}/>
-          <label className="brave-rewards-banner-logo" style={style.logoLabel} htmlFor="logoImageInput" >
-          </label>
-        </div>
-
-        <div className="brave-rewards-banner-background-image" style={style.backgroundImg}>
-        <input type="file" id="backgroundImageInput" style={style.imageInput} onChange={this.handleBackgroundImageUpload}  />
-        <label style={style.backgroundImageLabel} htmlFor="backgroundImageInput"></label>
-        </div>
-
-        <div className="brave-rewards-banner-content" style={style.bannerContent}>
-
-          <div className="brave-rewards-banner-content-social-links" style={style.socialLinks}>
-          <div className="brave-rewards-banner-content-social-links-youtube" style={style.socialLink}>
-          <YoutubeColorIcon/>
-          <input onChange={this.updateYoutube} className="brave-rewards-banner-content-social-links-youtube" style={style.socialLinkText} readOnly={this.props.mode !== 'Edit'} type="text" value={this.state.youtube}/>
-          </div>
-          <div className="brave-rewards-banner-content-social-links-twitter" style={style.socialLink}>
-          <TwitterColorIcon/>
-          <input onChange={this.updateTwitter} className="brave-rewards-banner-content-social-links-twitter" style={style.socialLinkText} readOnly={this.props.mode !== 'Edit'} type="text" value={this.state.twitter}/>
-          </div>
-          <div className="brave-rewards-banner-content-social-links-twitch" style={style.socialLink}>
-          <TwitchColorIcon/>
-          <input onChange={this.updateTwitch} className="brave-rewards-banner-content-social-links-twitch" style={style.socialLinkText} readOnly={this.props.mode !== 'Edit'} type="text" value={this.state.twitch}/>
-          </div>
-          </div>
-
-          <div className="brave-rewards-banner-content-explanatory-text" style={style.explanatoryText}>
-            <input style={style.explanatoryTitle} onChange={this.updateTitle} readOnly={this.props.mode !== 'Edit'} value={this.state.title} className="brave-rewards-banner-content-explanatory-text-headline" type='text'/>
-            <textarea style={style.explanatoryDescription} onChange={this.updateDescription} readOnly={this.props.mode !== 'Edit'} value={this.state.description} className="brave-rewards-banner-content-explanatory-text-headline" type='text'/>
-          </div>
-
-          <div className="brave-rewards-banner-content-donations" style={style.donations}>
-            <BatColorIcon style={style.batIcon}/>
-
-          <div className="brave-rewards-banner-content-donations-buttons" style={{height:'262px', marginLeft:'50px', paddingTop:'15px'}}>
-          <div className="brave-rewards-banner-content-donations-label" style={{fontSize:'15px'}}>Donation Amount</div>
-          <div className="brave-rewards-banner-content-donations-label" style={{fontSize:'12px'}}>Wallet Balance: 25 BAT</div>
-
-
-          <div style={style.donationsButtonContainer}>
-            <div className="brave-rewards-banner-content-donations-button" style={style.donationButton}>
-                <div style={style.donationsInput} className="brave-rewards-banner-content-explanatory-text-headline">{this.state.donationAmounts[0]} BAT</div>
-            </div>
-          <div className="brave-rewards-banner-content-donations-converted" style={style.donationsConverted}>{(this.state.donationAmounts[0] * this.props.conversionRate).toFixed(2)} {this.props.preferredCurrency}</div>
-          </div>
-          <div style={style.donationsButtonContainer}>
-            <div className="brave-rewards-banner-content-donations-button" style={style.donationButton}>
-                <div style={style.donationsInput} className="brave-rewards-banner-content-explanatory-text-headline">{this.state.donationAmounts[1]} BAT</div>
-            </div>
-          <div className="brave-rewards-banner-content-donations-converted" style={style.donationsConverted}>{(this.state.donationAmounts[1] * this.props.conversionRate).toFixed(2)} {this.props.preferredCurrency}</div>
-          </div>
-          <div style={style.donationsButtonContainer}>
-            <div className="brave-rewards-banner-content-donations-button" style={style.donationButton}>
-                <div style={style.donationsInput} className="brave-rewards-banner-content-explanatory-text-headline">{this.state.donationAmounts[2]} BAT</div>
-            </div>
-          <div className="brave-rewards-banner-content-donations-converted" style={style.donationsConverted}>{(this.state.donationAmounts[2] * this.props.conversionRate).toFixed(2)} {this.props.preferredCurrency}</div>
-          </div>
-
-          <div className="brave-rewards-banner-content-donations-checkbox" style={{paddingTop:'5px', marginLeft:'35px'}}>
-          <Checkbox value={'Checkbox values', {monthly : true}} size={'small'} multiple={false} disabled={false} type={'dark'}>
-            <div data-key='monthly'>Make this monthly</div>
-          </Checkbox>
-          </div>
-
-          </div>
-
-
-          </div>
-        </div>
-
-        <div className="brave-rewards-banner-bottom-bar" style={style.bottomBar}>
-        <div className="brave-rewards-banner-content-donations-send"
-        style={style.donationsCurrent}>
-        CURRENTLY DONATING BAT MONTHLY
-        </div>
-
-        <div className="brave-rewards-banner-content-donations-send"
-        style={style.donationsSend}>
-        SEND MY DONATION
-        </div>
-        </div>
-
-      </div>
-    );
-  }
-}
