+++ app/javascript/packs/brave_rewards_banner.style.jsx
@@ -1,8 +1,9 @@
 
   let bannerContent = {
     display: 'grid',
-    gridTemplateColumns: '3fr 5fr 4fr',
-    height: '262px'
+    gridTemplateColumns: '3fr 5fr 3fr',
+    backgroundColor: 'rgb(105, 111, 220',
+    height: '350px'
   }
 
   let donationButton = {
@@ -18,8 +19,8 @@
 
   let socialLinks = {
     backgroundColor: 'rgb(233, 240, 255)',
-    paddingTop: '100px',
-    paddingLeft:'40px'
+    paddingTop: '120px',
+    paddingLeft: '70px',
   }
 
   let socialLink = {
@@ -34,11 +35,14 @@
   let imageInput = {
     opacity:'0',
     position:'absolute',
+    visibility:'hidden'
   }
 
   let explanatoryText = {
     backgroundColor: 'rgb(233, 240, 255)',
-    paddingRight:'30px'
+    paddingTop:'30px',
+    paddingRight:'75px',
+    paddingLeft:'30px',
   }
 
   let donations = {
