--- app/javascript/packs/brave_rewards_banner_control_bar.jsx
@@ -1,15 +1,0 @@
-import React from 'react'
-
-export default class BraveRewardsBannerControlBar extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    return (
-      <div className="brave-rewards-banner-control-bar">
-      <p> Hi! Control bar! </p>
-      </div>
-    );
-  }
-}
