--- app/javascript/packs/brave_rewards_banner_intro.jsx
@@ -1,15 +1,0 @@
-import React from 'react'
-
-export default class BraveRewardsBannerIntro extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    return (
-      <div className="brave-rewards-banner-intro">
-      <p> Hi! Intro! </p>
-      </div>
-    );
-  }
-}
