+++ package.json
@@ -8,9 +8,9 @@
     "@types/react": "^16.4.7",
     "@types/react-dom": "^16.0.6",
     "babel-preset-react": "^6.24.1",
-    "brave-ui": "^0.23.0",
-    "clipboard": "^1.7.1",
+    "brave-ui": "^0.26.4",
     "chart.js": "^2.7.2",
+    "clipboard": "^1.7.1",
     "prop-types": "^15.6.2",
     "rails-ujs": "^5.1.4",
     "react": "^16.4.1",
